MDEV-21603: Long WHERE IN mariadb 10.3.22 crash
I_S tables and SHOW COMMANDS use special logic for creating and populating
temp.tables, e.g. table creation is delayed.
The code attempted to apply the same logic for the IN-to-SELECT temptable,
which resulted in an attempt to read from not-yet-created temptable and
crash.

Fix this by restricting use of I_S table logic only to I_S tables.