Deb: Ensure dch runs non-interactively so builds pass with new dch version
A recent version of dch (available in Ubuntu Hirsute and Debian Bullseye)
had a change in behaviour that it started prompting if the DEBEMAIL or
EMAIL variable as unset, asking for confirmation. We can't have anything
interactive in our build scripts, so prevent this prompt by giving
--controlmaint to the command, so it always uses the name and email from
the debian/control file and does not prompt anything.

The command-line argument has been around for a long time, so it is safe
to use on all Debian/Ubuntu builds we have.

See https://manpages.debian.org/jessie/devscripts/dch.1.en.html

Since MariaDB 10.2 is the oldest release we still release for Ubuntu Hisute
and Debian Bullseye, merge this on 10.2 and from there merge up to latest.