Deb: Drop dependency on libreadline-gplv2-dev on new distros
From Debian Bullseye/Ubuntu Hirsute onwards there is no longer any
libreadline-gplv2-dev available and it was replaced with libedit-dev in
commit 5cdf245d7e2ab339ad3dba0dbbb591ab80e0dad0.

This commit was however only applied on 10.5 and newer branches.
Since we still release MariaDB 10.2, 10.3 and 10.4 on new Debian and
Ubuntu releases, this specific change needs to be applied on the old
branches as well, but using a autobake-deb.sh modification that ensures
that we don't introduce new build dependencies on old MariaDB versions
on old Debian/Ubuntu releases.

MERGE NOTE: This change should only be merged from 10.2 on 10.3 and 10.4
and *not* on any newer branches.