MDEV-25583: Stop depending on empty transitional package dh-systemd
This amends the commit 058d93d47af79e650dc6a7abc107a400d8598e45 from
PR#1799 which lost this vital line in a rebase.

This change is intended to be applied for all MariaDB versions that will
be released for Debian Bullseye/Ubuntu Hirsute or newer. Since MariaDB 10.5
is already in Debian Bullseye/Ubuntu Hirsute, this commit only needs to
be applied on 10.5, 10.6 and newer.