MDEV-18959: Engine transaction recovery through persistent binlog
Implement handlerton functions to retrieve the least recoverable binlog
filename and position.

Within the iterator handlerton function each engine is requested to
provide last known binlog-offsets. Binlog file name is turned into
a numberical id. The engine's binlog-offset are compared to
compute the min and max which are also returned to the caller.