Make galera-4 Recommends instead of Depends in deb packaging
Galera is only necessary in MariaDB Server cluster setup. So, Galera
needs mariadb-server but mariadb-server doesn't.

Note that to be more close to what the Debian policy says, we should use
Suggests. But we use Recommends to be in sync with what we do in RPM
packaging.

  Recommends
  This declares a strong, but not absolute, dependency.
  The Recommends field should list packages that would be found together
  with this one in all but unusual installations.

  Suggests
  This is used to declare that one package may be more useful
  with one or more others. Using this field tells the packaging system
  and the user that the listed packages are related to this one and can
  perhaps enhance its usefulness, but that installing this one without
  them is perfectly reasonable.

See:
- https://github.com/MariaDB/server/commit/6bd94cf54274d54521ece1e50d534777122ff29e
- https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=988089
- https://www.debian.org/doc/debian-policy/ch-relationships.html#binary-dependencies-depends-recommends-suggests-enhances-pre-depends

Other Recommends packages moved are those used for the galera SST so are
also removed from Depends.