MDEV-25631: Crash in st_select_lex::mark_as_dependent with VIEW, aggregate and subquery
Name resolution code has checks like this one:

   if (thd->lex->in_sum_func &&
       thd->lex->in_sum_func->nest_level >= select->nest_level)
         ...

This fails to take into account the fact SELECT_LEX::nest_level is local
to each VIEW.

Adjust the check so that it only succeeds when the select and the aggregate
function being considered are from the same VIEW (or both from the top-level
query):  add this:

+  thd->lex->in_sum_func->nest_level_base == select->nest_level_base &&

Note: this patch only modifies one such check. There are many, should they all
be adjusted in the same way?