MDEV-16708: disabled the test main.sp to be executed with ps-protocol
This test has been disable by the reason that it contains multi-statements
that not supported in PS mode. Unfortunately, some of these multi-statements
can't be converted to a sequence of single-statements by the reason that
multi-statementness is a requirement for tests (e.g. tests for
the MySQL bugs #5307, #8762) therefore the whole test file will be
skipped in case it is run with --ps-protocol