MDEV-18959: a separate commit for refactoring in MDEV-21117 and the server base
recovery concerned binlog index files are enumerated in the order of
the binlog index file. Enumerated files are held in an introduced
 LOG_INFO::fname_hash
The binlog number identifies the file and is std::pair-ed with a
regular file offset to comprise a cross-file binlog offset which is
used for replication events identification and their mutual
binlog positioning comparison by the recovery procedure.

To merge with LOG_INFO refactoring commit.