Fixing optimizer code to generate better plans when optimizer_use_condition_selectivity=4
The Cause:
The condition selectivity for a table is calculated in “calculate_cond_selectivity_for_table”
function in sql/opt_range.cc. When calculating selectivity for indexes, the function iterates through
all the used keyparts and calculates selectivity. However, when we have extended keys enabled
and if a primary key part appear in the query in addition to a normal key, the selectivity
for the primary key part is not calculated.

Inside the “calculate_cond_selectivity_for_table” function, when calculate selectivity
for an index the code marks all the key parts of that key as processed.
When extended keys are enabled, the code also marks primary key parts
as processed. This causes primary key based condition selectivities to be not
calculated as the key parts are already marked as handled.

The Fix:
The code is modified to only consider user-defined keyparts when calculating the index selectivity.

Effect on existing test cases:
The following test case results were modified to match the output of the modified code.
        join_cache.result
        opt_trace.result