Introduce trx_sys_t::history_size_approx()
In btr0cur.cc there is potential for deadlock, if a rseg latch holder
(purge thread?) is waiting for an index page latch that the caller
of trx_sys.history_.... is holding while waiting for the rseg latch.

In some other places, we do not care about an exact count, and can use
an approximation that is computed without any latch protection.

Make also trx_sys_t::history_exists() a coarser approximation.