MDEV-20219: PURGE BINARY LOGS does not always purge
THIS COMMIT IS FOR DEMONSTRATION ONLY

PURGE BINARY LOGS (as intended) does not purge logs
that are open. When one thread is reading the logs, for
example via SHOW BINLOG EVENTS, then an issued
PURGE command will not delete the log while it is open
by SHOW. This commit provides a test/result to demonstrate
this behavior. With respect to MDEV-20219, the code inside
of the backup utility has a comment before purging stating
//be carefull with this step if you have slaves!
which leads me to believe this issue is due to the above
behavior.