MDEV-23857: replication master password length
Problem:
========
After MDEV-4013, the maximum length of
replication passwords was extended to 96
ASCII characters. After a restart, however,
slaves only read the first 41 characters of
MASTER_PASSWORD from the master.info file.
This lead to slaves unable to reconnect to
the master after a restart.

Solution:
========
After a slave restart, if a master.info file
is detected, use the full allowable length
of the password rather than 41 characters.

Reviewed By:
============
<TODO>