WIP: use of omp atomic for statisic_{inc,dec,add,sub}
TODO:

When parallel threads start to operate on a single
per session status variable there is more chance that inaccuracies
will occur. With the advent of atomics can maybe we use
"SAFE_STATISTICS" by default. The high content global variables
are being replaces with more optimium low contention counters.

thread_safe_{inc,dec,add,sub} could be removed as they are no
longer in use. But is this an API for plugins that needs to be
preserved?