Record iteration methods in JOIN_CACHE_BNL public
This commit makes public the methods to iterate the join
cache. This is useful for engine condition pushdown in
remote storage engines. In addition to push the condition,
we can also push data to match rows during JOIN.
This avoids full table scan and hence full table read
over the network.

Signed-off-by: Eduardo Berrocal <eduardo.berrocal@intel.com>