MDEV-26111 Spider handle ">=" as "=" in some cases
It is wrong to set `key_eq = TRUE` when `!use_both && tgt_key_part_map > 1`,
because it doesn't imply, at all, the original search condition is an equality.

In fact, if we specify the search condition `t1.b >= 1 AND t1.c > 0`,
then we get `end_key == NULL` and thus `!use_both`.