suppress galera error "Failed to report last committed"
according to MDEV-17550 it's informational, not fatal.
"last committed" is part of the certification index purge process.

depending on what tests and in what order are run it can be triggered
at unspecified times during the testing. If the test is happen to
shut down the server at this very time, the log will have:

[Warning] WSREP: Failed to report last committed XXXYYYZZZ, -77 (File descriptor in bad state)